requirements.txt
asgiref==3.3.1
Django==3.1.7
django-appconf==1.0.4
django-crispy-forms==1.11.1
django-discover-runner==1.0
django-role-permissions==3.1.1
django-select2==7.6.1
django-tempus-dominus==5.1.2.13
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
